add daemon_app::to_int checked-narrowing helper and apply at size/index boundaries
Replaces the 89 implicit qsizetype->int (and a few ->float/->double) narrowings flagged by
bugprone-narrowing-conversions with explicit conversions. New header-only daemon_app::to_int
(src/core/util/numeric.h) static_casts to int and asserts the value fits in int (full range, so
indexOf()==-1 and size()-1 on empty containers pass), catching a future out-of-range value instead
of silently truncating. qreal->float (QMatrix4x4) and ->double cases use inline static_cast.
The lint stays enabled; build + 48/48 ctest verified.

+++ src/core/fleet/mock_dashboard.cpp
@@ -114,10 +114,12 @@ QString MockDashboard::tokensToday() const {
         }
     }
     if (total >= 1000000) {
-        return QStringLiteral("%1M").arg(QString::number(total / 1000000.0, 'f', 2));
+        return QStringLiteral("%1M").arg(
+            QString::number(static_cast<double>(total) / 1000000.0, 'f', 2));
     }
     if (total >= 1000) {
-        return QStringLiteral("%1k").arg(QString::number(total / 1000.0, 'f', 1));
+        return QStringLiteral("%1k").arg(
+            QString::number(static_cast<double>(total) / 1000.0, 'f', 1));
     }
     return QString::number(total);
 }
